refactor(emitter): extract access.rs test module into submodule (§19)
Move the 1057-LOC `#[cfg(test)] mod tests` block out of
crates/tsz-emitter/src/emitter/expressions/access.rs into
access/tests.rs (a direct child module). access.rs drops 2554 -> 1499
LOC, bringing it under the 2000-line §19 limit, and the per-file
arch-guard ratchet is lowered 2554 -> 1499 in the same diff.

Pure code relocation: the test module uses `use crate::...` absolute
paths and has zero `super::`/`self::` references, so it resolves
identically after the move. All 44 relocated access tests pass.

AgentName: M1Opus48
